+++ tools/AutoLocalization/example/githook/post-commit
