+++ tools/AutoLocalization/example/githook/pre-commit
