+++ tools/AutoLocalization/main.py
@@ -1,4 +1,4 @@
-from src.cli import cli
+from src.auto_localization.cli import cli
 
 if __name__ == '__main__':
     cli()
