+++ tools/AutoLocalization/src/auto_localization/cli.py
@@ -5,14 +5,14 @@ from shutil import copy
 
 from dotenv import load_dotenv
 
-from src.file_loader.xaml_load import XamlParser
-from src.git import get_latest_file_content
+from src.auto_localization.xaml_load import XamlParser
+from src.auto_localization.git import get_latest_file_content
 
 logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
 logging.debug(os.path.abspath('.env'))
 load_dotenv(dotenv_path='.env')
-if os.path.exists('../.env'):
-    load_dotenv(dotenv_path='../.env')
+if os.path.exists('../../.env'):
+    load_dotenv(dotenv_path='../../.env')
 root_path = os.getenv("LOCALIZATION_PATH")
 assert root_path, "LOCALIZATION_PATH is not set"
 zh_cn_path = os.path.join(root_path, "zh-cn.xaml")
