+++ tools/AutoLocalization/src/auto_localization/git.py
@@ -29,4 +29,4 @@ def get_latest_file_content(file_path='./cli.py', encoding='utf-8'):
 
 
 if __name__ == '__main__':
-    print(get_latest_file_content('./cli.py'))
+    print(get_latest_file_content('cli.py'))
