+++ tools/AutoLocalization/src/auto_localization/translate.py
